+++ examples/cities/src/impl/solver_neighbourhood.cpp
@@ -28,12 +28,12 @@ bool __satisfies_constraints
 
 bool solver::can_remove_centre
 (
-	size_t loc_idx, const vector<role>& cities_served, vector<size_t>& new_locs
+	int loc_idx, const vector<role>& cities_served, vector<int>& new_locs
 )
 const
 {
-	vector<pair<double, size_t> > sorted_cities_served(n_cities);
-	for (size_t c = 0; c < n_cities; ++c) {
+	vector<pair<double, int> > sorted_cities_served(n_cities);
+	for (int c = 0; c < n_cities; ++c) {
 		if (cities_served[c] == none) {
 			sorted_cities_served[c].first = numeric_limits<double>::max();
 		}
@@ -55,10 +55,10 @@ const
 	vector<double> copy_cap_location = cap_location;
 	
 	// for every city ...
-	size_t c_idx = 0;
+	int c_idx = 0;
 	while (c_idx < n_cities and can_remove) {
 		
-		size_t city_idx = sorted_cities_served[c_idx].second;
+		int city_idx = sorted_cities_served[c_idx].second;
 		const point& city_pos = cities[city_idx].get_position();
 		new_locs[city_idx] = n_locations;
 		
@@ -67,10 +67,10 @@ const
 			
 			double max_gap = 0.0;	// gap = |cap_location + city_pop - centre_capacity|
 			double min_dist = numeric_limits<double>::max();	// distance between 'city_idx' and location 'l_it'
-			size_t min_loc = n_locations;
+			int min_loc = n_locations;
 			
 			// ... check every location ...
-			size_t l_it = 0;
+			int l_it = 0;
 			while (l_it < n_locations) {
 				
 				// ... that is not 'loc_idx', has a centre installed in it ...
@@ -82,7 +82,7 @@ const
 						double city_pop = cities[city_idx].get_population();	// added capacity
 						double dist_city_loc = city_pos.distance(locations[l_it]);	// distance between location and city
 						
-						size_t centre_idx = location_centre_type[l_it];
+						int centre_idx = location_centre_type[l_it];
 						double centre_cap = centres[centre_idx].get_capacity();// centre capacity
 						double wd = centres[centre_idx].get_working_distance();// centre's working distance
 						
@@ -138,9 +138,9 @@ const
 
 void solver::delete_centre
 (
-	size_t loc_idx,
+	int loc_idx,
 	const vector<role>& cities_served,
-	const vector<size_t>& new_locations,
+	const vector<int>& new_locations,
 	problem **neigh
 )
 const
@@ -155,9 +155,9 @@ const
 	//    2.1. Add the populations of cities served by location 'l'
 	//         to the new locations that serve them
 	//    2.2. set new locations serving the orphaned cities
-	for (size_t city_idx = 0; city_idx < n_cities; ++city_idx) {
+	for (int city_idx = 0; city_idx < n_cities; ++city_idx) {
 		
-		size_t new_idx = new_locations[city_idx];
+		int new_idx = new_locations[city_idx];
 		
 		if (cities_served[city_idx] == primary) {
 			new_sol->cap_location[new_idx] += cities[city_idx].get_population();
@@ -173,7 +173,7 @@ const
 	new_sol->cap_location[loc_idx] = 0.0;
 }
 
-void solver::replace_centre(size_t loc_idx, size_t centre_idx, problem **neigh) const
+void solver::replace_centre(int loc_idx, int centre_idx, problem **neigh) const
 {
 	*neigh = clone();
 	solver *new_sol = static_cast<solver *>(*neigh);
@@ -182,15 +182,15 @@ void solver::replace_centre(size_t loc_idx, size_t centre_idx, problem **neigh) 
 	new_sol->location_centre_type[loc_idx] = centre_idx;
 }
 
-bool solver::find_better_centre(size_t loc_idx, size_t& centre_idx) const {
-	size_t centre_installed = location_centre_type[loc_idx];
+bool solver::find_better_centre(int loc_idx, int& centre_idx) const {
+	int centre_installed = location_centre_type[loc_idx];
 	double instal_cost = centres[centre_installed].get_installation_cost();
 	
 	centre_idx = n_centres;
 	bool centre_found = false;
 	bool cheaper_centres_left = true;
 	
-	size_t ct_idx = 0;
+	int ct_idx = 0;
 	while (ct_idx < n_centres and not centre_found and cheaper_centres_left) {
 		
 		centre_idx = sorted_centres[ct_idx];
@@ -219,17 +219,17 @@ void solver::best_neighbour(pair<problem *, double>& bn, const local_search_poli
 	bool finish = false;
 	
 	// for all locations that have a centre installed
-	size_t loc_idx = 0;
+	int loc_idx = 0;
 	while (loc_idx < n_locations and not finish) {
 		
 		if (location_centre_type[loc_idx] != -1) {
 			// what cities is this location serving?
-			size_t n_cities_served;
+			int n_cities_served;
 			vector<role> cities_served(n_cities);
 			cities_served_by_location(loc_idx, cities_served, n_cities_served);
 			
 			// can we remove the centre? ...
-			vector<size_t> new_locations(n_cities);
+			vector<int> new_locations(n_cities);
 			bool can_remove = can_remove_centre(loc_idx, cities_served, new_locations);
 			
 			if (can_remove) {
@@ -257,7 +257,7 @@ void solver::best_neighbour(pair<problem *, double>& bn, const local_search_poli
 			}
 			else {
 				// ... we can not remove it ...
-				size_t cheaper_centre;
+				int cheaper_centre;
 				bool can_replace = find_better_centre(loc_idx, cheaper_centre);
 				
 				// ... but can we replace it with a cheaper one?
